requirements.txt
biopython
pandas
numpy
scikit-learn
flask
joblib
matplotlib
seaborn
python-dotenv
